+++ changelog.md
@@ -6,6 +6,7 @@
 * [fixed] Form behavior fixed for Missing taxon in media file
 * [fixed] Verification of media files
 * [changed] Most of the choice lists are sorted now
+* [added] Localities of Identity
 
 ## 02-2025